Replaced the standard djcelery 'task states' admin page with a version that uses the Backbone.js Celery class.
Not very useful yet, except as an example. When the 'task states'
admin page is loaded, the default table is emptied and replaced
with one populated by Backbone.js. The table is static at the
moment, although calling app.fetchTasks() will update the table.

In future I plan to build on the Backbone.js objects so polling
task status is supported.